refactor(connectors): organize add connector slices
