+++ apps/web/src/components/connectors/add-connector/manual-api-key-field.tsx
