+++ apps/web/src/components/connectors/add-connector/types.ts
@@ -1,0 +1,3 @@
+export type ConnectorConfigResult =
+  | { ok: true; value: Record<string, unknown> }
+  | { ok: false; message: string }
